% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{available_net_types}
\alias{available_net_types}
\title{Get Available Net Types}
\usage{
available_net_types()
}
\value{
Character vector of net type names.
}
\description{
Returns a character vector of available ITN (Insecticide-Treated Net) types.
}
\examples{
\dontrun{
net_types <- available_net_types()
print(net_types)
}

}
