% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emulator.R
\name{batch_predict_scenarios}
\alias{batch_predict_scenarios}
\title{Batch Predict Scenarios}
\usage{
batch_predict_scenarios(
  scenarios,
  predictor = "prevalence",
  batch_size = 1000L
)
}
\arguments{
\item{scenarios}{A data frame of scenarios.}

\item{predictor}{Character. Which prediction to generate.}

\item{batch_size}{Integer. Number of scenarios per batch.}
}
\value{
A data frame with predictions
}
\description{
Run predictions in batches for large numbers of scenarios.
}
