% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{clear_cache}
\alias{clear_cache}
\title{Clear Model Cache}
\usage{
clear_cache()
}
\value{
Invisible NULL
}
\description{
Clear all cached models from memory. This can be useful to free up
memory or to force models to be reloaded.
}
\examples{
\dontrun{
# Clear the cache
clear_cache()

# Models will be reloaded on next use
}

}
