% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emulator.R
\name{create_scenarios}
\alias{create_scenarios}
\title{Create Scenarios DataFrame}
\usage{
create_scenarios(
  eir,
  dn0_use,
  dn0_future,
  Q0,
  phi_bednets,
  seasonal,
  routine,
  itn_use,
  irs_use,
  itn_future,
  irs_future,
  lsm
)
}
\arguments{
\item{eir}{Numeric vector. Entomological Inoculation Rate.}

\item{dn0_use}{Numeric vector. Current net effectiveness.}

\item{dn0_future}{Numeric vector. Future net effectiveness.}

\item{Q0}{Numeric vector. Indoor biting proportion.}

\item{phi_bednets}{Numeric vector. Bednet usage proportion.}

\item{seasonal}{Numeric vector. Seasonality indicator.}

\item{routine}{Numeric vector. Routine treatment coverage.}

\item{itn_use}{Numeric vector. Current ITN coverage.}

\item{irs_use}{Numeric vector. Current IRS coverage.}

\item{itn_future}{Numeric vector. Future ITN coverage.}

\item{irs_future}{Numeric vector. Future IRS coverage.}

\item{lsm}{Numeric vector. LSM coverage.}
}
\value{
A data frame suitable for \code{run_malaria_emulator()}
}
\description{
Helper function to create a scenarios data frame for use with
\code{run_malaria_emulator()}.
}
