% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{define_bednet_types}
\alias{define_bednet_types}
\title{Define Bednet Types}
\usage{
define_bednet_types()
}
\value{
A list or data frame with bednet type definitions.
}
\description{
Get the definition/parameters for different bednet types.
}
