% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web_controller.R
\name{format_for_json}
\alias{format_for_json}
\title{Format Results for JSON}
\usage{
format_for_json(results)
}
\arguments{
\item{results}{A minter_results or mintweb_results object.}
}
\value{
A list formatted for JSON serialization
}
\description{
Format MINTe results for JSON serialization, suitable for web APIs.
}
