% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emulator.R
\name{generate_scenario_predictions}
\alias{generate_scenario_predictions}
\title{Generate Scenario Predictions}
\usage{
generate_scenario_predictions(scenarios, predictor = "prevalence")
}
\arguments{
\item{scenarios}{A data frame of scenarios.}

\item{predictor}{Character. Which prediction to generate.}
}
\value{
A data frame with predictions
}
\description{
Generate predictions for a set of scenarios using the loaded models.
}
