% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{minte_available}
\alias{minte_available}
\title{Check if minte Python module is available}
\usage{
minte_available()
}
\value{
Logical indicating if minte is available
}
\description{
Checks if the minte Python package is available and can be loaded.
This will trigger Python initialization if not already done.
}
