% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_prevalence}
\alias{plot_prevalence}
\title{Plot Prevalence Over Time (R native)}
\usage{
plot_prevalence(results, scenarios = NULL, title = "Prevalence Over Time")
}
\arguments{
\item{results}{A minter_results object or data frame with prevalence data.}

\item{scenarios}{Character vector. Which scenarios to plot. If NULL, plot all.}

\item{title}{Character. Plot title.}
}
\value{
A ggplot2 object
}
\description{
Create a ggplot2 visualization of prevalence over time. This is a native
R implementation that doesn't require Python plotting.
}
\examples{
\dontrun{
results <- run_minter_scenarios(...)
p <- plot_prevalence(results)
print(p)
}

}
