% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{preload_all_models}
\alias{preload_all_models}
\title{Preload All Models}
\usage{
preload_all_models()
}
\value{
Invisible NULL
}
\description{
Pre-load all neural network models into the cache for faster subsequent
predictions. This is useful when you want to minimize latency for the
first prediction call.
}
\examples{
\dontrun{
# Pre-load models at startup
preload_all_models()

# Now predictions will be faster
results <- run_minter_scenarios(...)
}

}
