% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controller.R
\name{print.minter_results}
\alias{print.minter_results}
\title{Print method for minter_results}
\usage{
\method{print}{minter_results}(x, ...)
}
\arguments{
\item{x}{A minter_results object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for minter_results
}
