% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web_controller.R
\name{print.mintweb_results}
\alias{print.mintweb_results}
\title{Print method for mintweb_results}
\usage{
\method{print}{mintweb_results}(x, ...)
}
\arguments{
\item{x}{A mintweb_results object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for mintweb_results
}
\keyword{internal}
