% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rminte-package.R
\docType{package}
\name{rminte-package}
\alias{rminte}
\alias{rminte-package}
\title{rminte: R Interface to Python's MINTe Package}
\description{
An R interface to the Python 'minte' package, providing neural network-based
malaria scenario predictions for evaluating intervention strategies including
ITNs (Insecticide-Treated Nets), IRS (Indoor Residual Spraying), and LSM
(Larval Source Management).
}
\section{Main Functions}{

\itemize{
\item \code{\link{run_minter_scenarios}}: Main entry point for running scenarios
\item \code{\link{run_mintweb_controller}}: Simplified web interface controller
\item \code{\link{run_malaria_emulator}}: Direct emulator interface
\item \code{\link{create_scenario_plots}}: Create visualizations
}
}

\section{Utility Functions}{

\itemize{
\item \code{\link{calculate_overall_dn0}}: Calculate net effectiveness
\item \code{\link{available_net_types}}: Get supported net types
\item \code{\link{preload_all_models}}: Pre-load models into cache
\item \code{\link{clear_cache}}: Clear model cache
}
}

\section{Setup}{

\itemize{
\item \code{\link{minte_available}}: Check if minte is available
\item \code{\link{install_minte}}: Install minte Python package
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mrc-ide/rminte}
  \item \url{https://mrc-ide.github.io/rminte/}
  \item Report bugs at \url{https://github.com/mrc-ide/rminte/issues}
}

}
\author{
\strong{Maintainer}: OJ Watson \email{o.watson15@imperial.ac.uk} (\href{https://orcid.org/0000-0003-2374-0741}{ORCID})

}
\keyword{internal}
