% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emulator.R
\name{predict_full_sequence}
\alias{predict_full_sequence}
\title{Predict Full Sequence}
\usage{
predict_full_sequence(scenarios, predictor = "prevalence")
}
\arguments{
\item{scenarios}{A data frame of scenarios.}

\item{predictor}{Character. Which prediction to generate.}
}
\value{
Predictions array
}
\description{
Generate full sequence predictions for scenarios.
}
